% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGRanges.R
\name{readGRanges}
\alias{readGRanges}
\title{Read a sequence region file into a GRanges object}
\usage{
readGRanges(x, seqinfo = GenomeInfoDb::seqinfo(x), ...)
}
\arguments{
\item{x}{File name to read into GRanges object}

\item{seqinfo}{Seqinfo object with metadata about the genome}
}
\value{
A GRanges object containing all the reads of the sequence file
}
\description{
Read a sequence region file into a GRanges object
}
